- Remove Listtile from MedicationScreen, replace for MedicationCard. - Implement update Medicament. - Implement view and register data without network, sync when have network

+++ lib/features/medicines/data/repositories/medicament_repository_impl.dart
@@ -1,3 +1,4 @@
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:pasti_track/core/config.dart';
 import 'package:pasti_track/core/errors/failures.dart';
 import 'package:pasti_track/core/helper/app_logger.dart';
@@ -9,16 +10,20 @@ import 'package:pasti_track/features/medicines/domain/repositories/medicament_re
 class MedicamentRepositoryImpl implements MedicamentRepository {
   final MedicamentLocalDataSource localDB;
   final MedicamentRemoteDataSource remoteDB;
+  final Connectivity _connectivity = Connectivity();
 
   MedicamentRepositoryImpl(
     this.localDB,
     this.remoteDB,
   );
 
+  Future<bool> isConnected() async {
+    var connectivityResult = await _connectivity.checkConnectivity();
+    return connectivityResult.first != ConnectivityResult.none;
+  }
+
   @override
   Future<List<Medicament>> getMedications() async {
-    AppLogger.p(" Medicament", getMedications);
-
     try {
       final localMedications = await localDB.getMedications();
       AppLogger.p("Medicament getMedications", localMedications);
@@ -32,10 +37,10 @@ class MedicamentRepositoryImpl implements MedicamentRepository {
   @override
   Future<int> addMedicament(Medicament medicament) async {
     try {
-      final result1 = await localDB.addMedicament(medicament);
-      await remoteDB.addMedicament(medicament);
+      final result = await localDB.addMedicament(medicament);
+      if (await isConnected()) await remoteDB.addMedicament(medicament);
       AppLogger.p("Medicament", "addMedicament ");
-      return result1;
+      return result;
     } catch (e) {
       AppLogger.p("Catch Medicament", "addMedicament ${e.toString()}");
       throw Failure(AppString.errorWhenCreate(AppString.medicament));
@@ -43,28 +48,69 @@ class MedicamentRepositoryImpl implements MedicamentRepository {
   }
 
   @override
+  Future<int> updateMedicament(Medicament medicament) async {
+    try {
+      final result = await localDB.updateMedicament(medicament);
+      if (await isConnected()) await remoteDB.updateMedicament(medicament);
+      AppLogger.p("Medicament", "updateMedicament $result");
+      return result;
+    } catch (e) {
+      AppLogger.p("Catch Medicament", "updateMedicament ${e.toString()}");
+      throw Failure(AppString.errorWhenUpdate(AppString.medicament));
+    }
+  }
+
+  @override
   Future<int> deleteMedicament(String id) async {
     try {
-      final result = await localDB.deleteMedicament(id);
-      await remoteDB.deleteMedicament(id);
-      AppLogger.p("Medicament", "deleteMedicament $result");
-      return result;
+      if (await isConnected()) {
+        final result = await localDB.deleteMedicament(id);
+        await remoteDB.deleteMedicament(id);
+        AppLogger.p("Medicament", "deleteMedicament $result");
+        return result;
+      } else {
+        throw Failure(AppString.errorWhenDelete(
+            AppString.canNotBeActionCheckYourConnectivityTryAgain));
+      }
+    } on Failure catch (e) {
+      AppLogger.p("Catch Medicament Failure", "deleteMedicament ${e.message}");
+      throw Failure(e.message);
     } catch (e) {
       AppLogger.p("Catch Medicament", "deleteMedicament ${e.toString()}");
       throw Failure(AppString.errorWhenDelete(AppString.medicament));
     }
   }
 
-  @override
-  Future<int> updateMedicament(Medicament medicament) async {
-    try {
-      final result = await localDB.updateMedicament(medicament);
-      await remoteDB.updateMedicament(medicament);
-      AppLogger.p("Medicament", "updateMedicament $result");
-      return result;
-    } catch (e) {
-      AppLogger.p("Catch Medicament", "updateMedicament ${e.toString()}");
-      throw Failure(AppString.errorWhenUpdate(AppString.medicament));
+  Future<void> syncData() async {
+    if (await isConnected()) {
+      List<Medicament> localMeds = await getMedications();
+      List<Medicament> remoteMeds = await remoteDB.getMedications();
+
+      Map<String, Medicament> remoteMedMap = {
+        for (var med in remoteMeds) med.medicineId: med
+      };
+
+      for (var localMed in localMeds) {
+        if (remoteMedMap.containsKey(localMed.medicineId)) {
+          Medicament remoteMed = remoteMedMap[localMed.medicineId]!;
+
+          DateTime localDate = DateTime.parse(localMed.dateUpdated);
+          DateTime remoteDate = DateTime.parse(remoteMed.dateUpdated);
+
+          if (localDate.isAfter(remoteDate)) {
+            await remoteDB.updateMedicament(localMed);
+          } else if (remoteDate.isAfter(localDate)) {
+            await localDB.updateMedicament(remoteMed);
+          }
+          remoteMedMap.remove(localMed.medicineId);
+        } else {
+          await remoteDB.addMedicament(localMed);
+        }
+      }
+
+      for (var med in remoteMedMap.values) {
+        await localDB.addMedicament(med);
+      }
     }
   }
 }
